Remove "input" as a qualifier to "argument"
Fixes #23327

A reader was concerned that these attributes could not be applied to arguments that might have the `out` qualifier. The word is redundant, so remove it from every location where "input" qualified either "parameter" or "argument". In other cases, where "input" was used without "argument", replace "input" with "argument". That's more correct and more consistent.