Update distributed tracing docs (#23313)
* Update distributed tracing docs

We are about to announce support for OpenTelemetry .NET and there is
likely to be an influx of developers reading these docs. They are
already the #1 link for "distributed tracing" on bing.com so they are
probably getting above average traffic/scrutiny too.

- Reorganized the content into a landing page, conceptual docs,
instrumentation walkthroughs and collection walkthroughs
- Added some more intro and conceptual information assuming that many
developers are completely new to distributed tracing and have no
frame of reference.
- Removed the example of doing distributed tracing with DiagnosticSource
We might want to restore this in the future but hopefully it represents
a niche usage scenario that we are migrating away from. Without context
indicating when to use it it is likely to confuse new developers about
recommended patterns.
- Updated the examples to ensure they start with copy-pastable code
that is fully runnable.
- Added links for collecting distributed tracing instrumentation with
Application Insights

There are many oportunities for further improvement that we may
want to triage:
- Add examples and pictures of more realistic collected distributed
trace information to help developers internalize why it might be useful
to them.
- Add a diagnostic guide showing how distributed tracing assists in
resolving a realistic production issue.
- Ensure the getting started guides for collection transfer smoothly to
remote pages in the AppInsights and OpenTelemetry docs
  - There was a spot I wanted to link to supported OpenTelemetry
     exporters but there is no anchor that goes solely to that
     information.
- Flesh out the instrumentation guidance to answer more questions
developers may encounter such as:
  - naming conventions for Activity
  - when to use DisplayName
  - how to log exceptions
  - how to propagate IDs to and from network protocols
  - how to migrate pre-existing Activity instrumentation to use
     ActivitySource
- Add performance analysis and guidance to reassure developers that
  Activity instrumentation + OpenTelemetry are serious about high
  performance workloads.
- Discuss more conceptual topics and/or go into greater depth about
some areas:
  - Hierarchical vs. W3C ID.
  - Format of a W3C ID
  - Describe sampling options in ActivityListener
  - Explain why some activities are uncapturable with ActivityListener
     and require workarounds with DiagnosticListener.
 - Roadmap for handling issues around custom propagators, supporting
protocol updates, and suppressing undesired automatic Activity creation