Update how-to-write-csharp-analyzer-code-fix.md (#23500)
Build Unit Tests doesn't remove the test methods created by the template, which will no longer pass based on the changes to the analyzer and code fix.  Then, the tutorial gives a single test method to add but asks the user to run "these two tests" -- fixed to refer to only the single, added test.