Fix RegexOptions.IgnorePatternWhitespace # to EOL
The documentation currently states that

> All text in the regular expression pattern from the # character to the
> end of the string is interpreted as a comment.

when `RegexOptions.IgnorePatternWhitespace` is in effect.  This omits
the fact that `#` behaves as a line comment when the pattern contains
multiple lines.  It causes the text from `#` to the end of the line,
denoted by `\n`, to be interpreted as a comment.  For example:

```csharp
using System.Text.RegularExpressions;

Console.WriteLine(Regex.Match(
    "Hello World",
    "Hello # Line Comment\n\\sWorld",
    RegexOptions.IgnorePatternWhitespace));
```

Prints "Hello World", rather than "Hello" as it does if `\n` is removed
or replaced by `\r`.

This commit adds this behavior to the documentation.

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>