Record types can seal ToString in C# 10. (#24957)
* Add information on sealed ToString

Fixes #24077

Beginning in C# 10.0, records may declare a `ToString` method as `sealed`. This prevents derived records from creating an override. Effectively, that means the `ToString` output will not include the runtime type information. (All members and values are displayed, because derived records will still have a PrintMembers method generated.)

* add note on ToString.

* fix a build warning introduced

This was introduced in an earlier PR

* Update docs/csharp/language-reference/builtin-types/record.md

Co-authored-by: Youssef Victor <youssefvictor00@gmail.com>

* Apply suggestions from code review

Co-authored-by: Tom Dykstra <tdykstra@microsoft.com>

Co-authored-by: Youssef Victor <youssefvictor00@gmail.com>
Co-authored-by: Tom Dykstra <tdykstra@microsoft.com>