[WIP] Add new metric APIs
We added new metric APIs in .NET 6 and we need to tell customers about them.
These initial pages disambiguate the API from pre-existing APIs and help
users pick.

I plan to add more content that lets devs understand and use
the new APIs in a new page that is parallel to the EventCounters page, but
I haven't gotten to that part yet.