Discussion from runtime/issues/12506
https://github.com/dotnet/runtime/issues/12506#issuecomment-895276247

This article currently covers the verbose and (arguably, less common) way of using dotnet-trace to debug assembly loading issues, as most assembly loading issues likely occur when the generic IHostBuilder is initially called.

I debated rewriting the whole article to emphasize this use case, but I tempered that to simply get this documented with what I feel is the natural way to trace assembly loading events.

For the content, I elected to mostly use the docs @vitek-karas team has already written, here: https://github.com/dotnet/diagnostics/blob/main/documentation/dotnet-trace-instructions.md#using-dotnet-trace-to-launch-a-child-process-and-trace-it-from-startup

I realize copy-paste is not great, but I felt it was important to capture in this article the full flow of calling a new process.  I also fixed some minor stuff like kbd escape tags.