edit pass on get-started material
Fixes #25014

Many of the concerns described in the issue had been addressed in previous edit passes. (This issue was entered in a different feedback system, and was more recently moved to our GitHub repo). The edits addressed the remaining issues, must of which are concerned with including definitions for terms as they are introduced.