Use `EnableDynamicLoading` for plugins
In .NET 6 SDK doesn't produce `.runtimeconfig.dev.json` by default and as such plugins with NuGet dependencies won't run with just `dotnet build` as they don't copy their dependencies to output.

Adding `EnableDynamicLoading=true` to a plugin project to fix this. Technically it's not necessary, and just `CopyLocalLockFileAssemblies=true` would be enough, but I think it's just really weird. Also it's much easier to explain: "I want to load this project dynamically at runtime, so I'm going to mark it with `EnableDynamicLoading=true`".

Modified the section which adressed this issue previous - mostly simplifying it.