Deprecate and remove the coreclr.h based hosting
Hosting .NET through `coreclr.h` has been obsolete since .NET Core 3.0 and should not be used.

All the versions which needed it (.NET Core 2.2 and older) are not out of support.

Having the `coreclr.h` hosting available in the docs and samples is confusing and may lead people to actually use it for newer runtimes which can lead to unexpected results.

This change removes the sample code and modifies the hosting guide to only mention one hosting API. I added a note about the older runtimes needing to use `coreclr.h` but intentionally didn't include any details on that topic.