"Code-style naming rules" amended
The "Code-style naming rules" MS docs page is currently hard to grasp:

There is a mistake in the heading hierarchy, making it difficult to find content.

The word "property" is used for distinct entities, making it hard to understand which entity a piece of information applies to.

The sequence of sections isn't congruent with the information to convey.

The tiny "Required" table column information is lost when put behind the huge "Allowed values" column.
<br/>

**So, I tried to fix the content to make it easier to understand.**

I did the following:

- Heading hierarchy fixed
- "property" replaced with "component", or "rule", where applicable
- Sections semantically sorted
- Additional pieces of information added to tables
- "Required" table column moved left to "Allowed values" column, so the information won't be lost
- Additional example provided, to demonstrate usage of `required_prefix`