--version-suffix example uses incorrect naming
semver 2.0 should use `rc.1" not "rc1", as per conversation I had with @joelverhagen where I was using it incorrectly.  I believe the example in this article is likely the root cause as to why I thought `rc1` was correct, so it would be great to update and fix this. See discussion here https://github.com/NuGet/NuGetGallery/issues/8325