CS1617: Clarify valid language versions (#20259)
Clarify for CS1617 that the list of valid language versions depends on
the .NET version that is being used to compile the application. Newer
language versions will only be available by .NET versions that came
after the language version was introduced. The displayed list only
shows the versions for the current version of .NET.