Update code style for private fields (#26610)
Update example snippets to follow the code style for private and static
fields, and use an underscore or a `s_` prefix.

This commit only changes the styling in those examples where the field
is not the primary focus of the documentation page. This avoids
over-focusing on the styling in favor of keeping the examples succinct,
and also helps not to distract from the actual topic. Otherwise, many
examples, like public fields, couldn’t be explained because the style
guide discourages them.

- Use underscrore prefix for private fields.
- Always specify visibility modifiers.
- Use `s_` prefix for static fields.
- Avoid public field in `override` docs and use a public property
  `Employee.Name` instead.
- Mention abstract properties in `abstract` docs and move explanatory
  comments above lines.