Clarify how to set SCH_USE_STRONG_CRYPTO flag (#26604)
The text makes is sound, like the Schannel flag can be set directly in registry, while the truth is it can be only passed via APIs to Schannel -- .NET Framework needs to be told to pass it via registry or app context. Adding links to appropriate sections in the doc describing how to do that.