Fix the build for VS2022 by altering the arguments to ArgumentException
Why: Currently this project does not build for users of VS2022 because of CA2208.
Fix page mentioned below suggests two solutions. I implemented the second
because that seems to be most in line the original intend: only pass in the name
of the violating parameter.
See: https://docs.microsoft.com/nl-nl/dotnet/fundamentals/code-analysis/quality-rules/ca2208#how-to-fix-violations