Rewrite 'Avoid multiple asserts' to 'Avoid multiple acts' #27198 (#27200)
* Fix the build for VS2022 by altering the arguments to ArgumentException

Why: Currently this project does not build for users of VS2022 because of CA2208.

* Rewrite samples for Avoid multiple acts

Why: The previous samples did not make a distinction
between the act and assert, now that we are talking
about avoiding acts it become more important to
make the acts more explicit.

* Rewrite to 'Avoid multiple acts'

Why: The problem shows in the multiple asserts
but is caused by the multiple acts. When rewritten as
using multiple acts the exception to the rule can be
removed.