Update choosing-between-anonymous-and-tuple.md
Line 89: changed "... however; there are slight differences..." to "...however, there are slight differences..." 

Just a simple grammar mistake - using a semicolon instead of a comma. Although, the semicolon could stay if the position of" however,..." was moved after it.

Hopefully this is not too nit-picky and a waste of time; I saw the mistake and was excited by the possibility of contributing to the Microsoft Docs.