Update tutorials to .NET 6
Fixes #27896

Also, in this PR, update all the tutorials in the Get Started and Tutorials section to .NET 6. That includes using top-level statements, implicit usings, and following other conventions.

Because I changed so much of the included code, I updated the date on the associated articles even though the text didn't change as much.