Give developers a reason to switch
"open-source", "cross-platform" and "interoperable" are basic necessities for any programming language so I have demoted them to the second paragraph.
"Succinct" may be accurate when one considers an entire program, but is not obvious when looking at say a C# lambda vs an F# one. "Robust" is too vague to be meaningful and "performant" is just another basic necessity. 

"Your focus remains on your problem domain, rather than the details of programming" and "data-oriented, where code involves transforming data with functions" may be accurate but are not going to cut through to the target audience of developers contemplating a switch to F#.

What are the banes of non-F# programmer's lives? 1. Clutter 2. Having to write documentation and 3. Having to read documentation that may or may not accurately describe the code.

F# solves developers' biggest headaches and this should be front and centre in explaining "What is F#".