If/then/else is not so different
Maybe it's just me, but when I first started with F#, I was scared to write code that said if FileIsOk then WriteFile because I was intimidated by documentation that made a clear distinction between F#'s cool new if-then-else expression and the discredited if-then-else statements of lesser languages. Scott Wlaschin even went as far as to say use of if-then-else is strongly discouraged.
The changes I made to this page are designed to help programmers coming from imperative languages relax a bit and feel good about themselves again.