Document Diagnostic Port
Fixes https://github.com/dotnet/diagnostics/issues/2900

We added the diagnostic port in 3.1 and expanded on it in 5.0, but never
documented it aside from tangential references in some tools that used it.
This change describes it directly and adds some detail that wasn't mentioned
elsewhere. I also tried to improve docs on some related environment variables
that misleadingly claimed scenarios were only for dotnet-monitor or only for
Mono.