Publish from the draft v7 branch (#29100)
The committee has finished the C# 6 standard, and submitted the draft to ECMA. It's now turning its focus to the C# 7.x release(s).

The PR changes the branch from which we publish the standard, but doesn't do any further work. The content is at v6, as we haven't started accepting any V7 PRs.