Lowered verbosity
Lowered verbosity on the sections for expression-bodied member and auto-implemented properties. This article should be about init, and how it can be used in those contexts, but I believe there is no need to further explain the merits of auto-implemented properties, or the capabilities of the C# compiler, for example, those have their own page. The fact that they suport init and how should be enough, and much less confusing imo.