Fixes #28385 (#29925)
* Fixes #28385

Added clarification on the usefulness of the init keyword. Added code snippet to further exemplify it.

* More relevant property to the use of init

* Changed property to more relevant use of init

* Update InitExample3.cs

* Changed property to more relevant use of init

* Lowered verbosity

Lowered verbosity on the sections for expression-bodied member and auto-implemented properties. This article should be about init, and how it can be used in those contexts, but I believe there is no need to further explain the merits of auto-implemented properties, or the capabilities of the C# compiler, for example, those have their own page. The fact that they suport init and how should be enough, and much less confusing imo.

* Edit to non working code

* Update InitExample3.cs

* Apply suggestions from code review

Co-authored-by: Bill Wagner <wiwagn@microsoft.com>