Removed ISiloHost reference in Orleans Tutorial (#29935)
Orleans recently switched from a specialist SiloHostBuilder, which would construct a SiloHost, to use the generic host builder (Microsoft.Extensions.Hosting.HostBuilder), which constructs an IHost. The code sample does not compile as-is, which may dissuade newcomers to Orleans. this change advises to add the Microsoft.Extensions.Hosting nuget package and updates the code sample to include the using statement and changes the ISiloHost return type to IHost. Built and tested locally successfully.