Promote use of GenerateDocumentationFile over DocumentationFile
For a documentation file to be correctly picked up by the IDE, it must have a specific name and be in a specific location. It is easier and more maintainable to opt into documentation file generation with a boolean value than with a file path. For this reason, we should promote use of `GenerateDocumentationFile`, with only a passing reference to the older `DocumentationFile` property for rare case that the default behaviour should be overridden.