Fix placement of summary paragraph (#30247)
When the section "List patterns" was added in https://github.com/dotnet/docs/commit/b181812600f3a1efe8a57fefd7dee80a469d9f93, it was placed after the paragraph that was intended to be at the end of the article. This commit puts the summary paragraph back at the end where it was intended to be.