Update the documentation for PublishRelease and PackRelease
For .NET 7.0.100-rc1 we have enabled `PublishRelease` to work inside of a `.csproj` instead of forcing it to be put in `directory.build.props`. We've also introduced an environment variable to enable it for solutions. (The reason for the environment variable is so users can opt-in and or out of the performance cost of this new feature.)

See the relevant PR from the SDK here: https://github.com/dotnet/sdk/pull/26808
Old commit: https://github.com/dotnet/docs/pull/29996