Add Dotnet Workload command page.
I based this off of the `dotnet` command page. We support `dotnet workload --info` and needed to have this page to document it. It should show up as the primary header for workloads most likely, not sure how to make that change reflected in the main docs page. 

Since this may be the first thing people see searching up workloads, it should be well polished. Requesting a review from the SDK team to review.