Add full list of warnings as header
I'm seeing higher bounce-back on this article than I'd like.

As an experiment, I'm adding a table at the top of the article with all the warnings covered in this article, and links to the section where that concept is covered.

The hypothesis is that readers can quickly click on the warning # to navigate to the correct section quickly.