improve SEO on statements (#32704)
* improve SEO on statements

In all the statements pages, edit the title, description, and opening paragraph to increase SEO on the statement keywords.

* Update all the operator pages

* Improve SEO for tokens

The special characters `@` and `$` are hard to search for. Include the names in the title and descriptions to improve possible search terms.

* fix warnings

* restore some lists for readability

In some of the cases, the lists of statements or operators is a much more readable format than the prose. In those cases, expand the opening paragraph to include the keywords for likely search terms, and restore the list.