Duplicate line removed
The current line 61 `Console.WriteLine(string.Join(" ", numbers));  // output: 5 0` was repeated on line 64. I removed 64. 

The line 63 was `numbers.Add(a ??= 0);` i changed it to `numbers.Add(a ??= 3);` because in my opinion an int > 0 conveys clearly that it is not `null`. 
On a side note in dutch / flemish and german zero `0` is called  `Nul` or `Null` which is pretty close to `null`