Add 3rd party dependencies guidance
Fixes #32861

The dotnet content team has applied a consistent set of rules regarding recommendations and dependencies in our content. What we haven’t done is state those rules publicly. That resulted in misunderstandings after we removed a previous recommendation. We want to resolve that by publishing the rules we use.