Fix issues.
**This commit should be reviewed carefully**

There are three issues open on this article. Fix two of them first:

Fixes #26446: Clarify where to find the HttClient class.
Fixes ##27720: Add call to `ToArray` to force eager evaluation of a LINQ query. Explain why this was added.

The other issue may require a new article to limit the number of concurrent tasks.