Create async section (#33921)
* Fix issues.

**This commit should be reviewed carefully**

There are three issues open on this article. Fix two of them first:

Fixes #26446: Clarify where to find the HttClient class.
Fixes ##27720: Add call to `ToArray` to force eager evaluation of a LINQ query. Explain why this was added.

The other issue may require a new article to limit the number of concurrent tasks.

* Move async article

Move the scenarios overview to the new section.

* Fix issues before moving file

**This commit should be reviewed**

There are six issues open on this file.  Fix the following:

Fixes #20680 Clarify that starting a task and immediately awaying doesn't result in any faster results.
Fixes #33641: Explain why the finished task must also be awaited.

* Move existing async overview article.

Move this article under the new async section.

* move async task model

* Move async return types article.

move samples

* Fix open issues on this page.

**This commit should be reviewed**

This page has one open issue.

Fixes #32782:  Add a clarification to the example code to note that the while loop will complete.

* move article on processing multiple tasks as they finish

* Move async file access article and sample.

* Fix open issues on this article

**This commit should be reviewed**

The only open issue is #27061

Update the sample, and add text to explain the update.

* Move cancellation articles

Move the tutorials on cancelling a running task.

* Fix open issues on tutorial

**This commit should be reviewed**

There's only one open issue on this tutorial:

* Move async streams tutorial

* fix warnings

restore and move images

These images were under the old C# folder. They are only used in the VB docs, so restore them here.

* update breadcrumb, redirections and metadata.

** This commit should be reviewed**

* Apply suggestions from code review

Co-authored-by: David Pine <david.pine@microsoft.com>

* Respond to feedback. Remove specific .NET 6.0 comments.

---------

Co-authored-by: David Pine <david.pine@microsoft.com>