Create native interop section (#34132)
* Move and consolidate overview

The previous "index" article was an old landing page that had only an introductory paragraph and then links to the additional articles in the section. Consolidate the opening paragraph with the overview.

* Move office How to article

Snippets haven't moved until I get clarity on sample story.

* Move indexed properties article

* move how to play a wave file

* Split and move the office programming walkthrough

Split this article in 2 while moving it. This will make it easier to find for VB developers looking for the VB version. And, the C# version reads more cleanly.

* Move example COM class

* remove duplicate redirection

* build warnings

* one more build warning

* move sample for example COM class

* move office walkthrough sample

* add PInvoke samples

Move the PInvoke sample to the new folder.

* move indexed sample

* fix the folder

* update for expected errors

* move final sample

* fix a build warning

* Move Office article on named / optional arguments.

* fix warnings

* more build issues

* one more build issue

* one more build error

* move using type dynamic

* move and duplicate dynamic walkthrough

* fix build error

* fix build warnings

* fix relative links to moved files.

* code review

* proofread and style updates.

* fix build warnings

* restore missing snippet

* respond to feedback

* ignore one more error

* test