Update breaking-changes.md
changed "almost any change in behavior could result breaking someone" to be "almost any change in behavior could cause a logic error for a consumer".
I very much doubt the author intended to say "breaking someone", maybe "breaking something" and that's why I added 'for a consumer'. Personally I prefer "almost any change in behavior could cause a logic error" but will leave it changed as-is to keep as close to what I believe is the authors intent.