fix redirection and breadcrumb (#34624)
Fix the breadcrumb for the C# "advanced topics" section. There isn't an index in that folder, so defer to the first sub-section and it's overview article.

Fixes #34604

Also, fix the redirection to this same article: The destination must include the training "/" to redirect to the (default) `index` in that folder.