Add information on interceptors
Add what's new and error / warning information for Interceptors.

Fixes #35878

This feature is experimental, so for details I'm linking to the feature spec in the roslyn repo. The feature may change. When it's finalized, I'll add more details on it in docs, and remove the warnings about its experimental status.