Update formatting.md
Disroot the conception that "vanity" is the reason code may use whitespace alignment. It may be for other reasons (readability).

Despite the guideline discourages formatting that relies on length of identifiers, and that some diff tools will show whitespace changes as if the whole line changed, I think it is better to not carry judgement through "vanity" adjective.