Fix naming conventions
Fixes #30626: Clarify (again) that these are our guidelines, not yours. Point out that it's not the VS default, but a configuration option.
Fixes #30642: Again, our style.
Fixes #30799: Change constant style from ALL_CAPS to PascalCase to match runtime repo.
Fixes #33959: Update variable names so delegate types are PascalCased and instances of a delegate are camelCase. Add clarifying text for the same.