Combine and add expression tree restrictions
Contributes to #36356

One of the new compiler errors added for collection expressions is that collection expressions are not allowed in expression trees.

There are many features in the language that can't be used in expression trees. While adding this new error create a page to handle all the expression tree restrictions.