Add customer feedback GitHub template. (#37043)
This is the first pass at the new issue template for feedback from the live site control.

Contributes to #37025

In this first PR, I've tried to replicate the information in the current template as best as I can. I've got a question to the team to see if ms.author and author metadata is supported yet.