Fix casing in named placeholder in sample Worker.cs
This fixes invalid camelCase in the logging code sample `Worker.cs` in favour of PascalCase which should be used.

Details are explained in [the CS1727](https://learn.microsoft.com/en-us/dotnet/fundamentals/code-analysis/quality-rules/ca1727).

The whole page is inconsistent. The proper casing is used at other samples, but not in `Worker.cs`. This change removes this inconsistency.

The current state is confusing for less experienced .NET developers.