Reflect in the documentation that the LOHThreshold has to be less than region_size
We recently discovered that we cannot let the user to allocate arbitrarily large objects in the small object heap, therefore we implement a change (https://github.com/dotnet/runtime/pull/95485) in the runtime and this change in the doc reflects that.

The actual cap is the `gc_region_size`. It is an implementation detail and it changes according to different user/machine configurations. Therefore, instead of specifying the maximum, we just said it could be capped.