Edit pass for LINQ "get started" section. (#38723)
* Fix Open issues

Fixes #28863 - Add a note to include a using for `System.Linq`, unless you're project includes implicit usings. Add links to both.

Fixes #36684 - The text in the article used the phrase "between 3 and 7", but the sample was less than 3 or greater than 7. Fix the code to match the text.

* move files

Move all MD files, associated snippets, and associated media for the "get started with LINQ" section into that section.

Move test project to new location.

Delete orphaned snippet projects.

* refactor sample source

Make sure all snippets are called in the main program. Remove any unused code in the snippets for this folder.

* proofread and edit.

* replace redirects.

* replace absolute links with relative links

* remove redundant links

* Apply suggestions from code review

Co-authored-by: Tom Dykstra <tdykstra@microsoft.com>

* respond to feedback

---------

Co-authored-by: Tom Dykstra <tdykstra@microsoft.com>