edit remove-runtime-sdk-versions.md: os-macos version: remove reference to Linux distribution
This page has a toggle choose from Windows, Linux, and macOS.

When I choose the option for macOS, I see in the 'Uninstall .NET' section a reference to a different operating system: Linux

> These directories may vary depending on your Linux distribution.

This commit replaces 'Linux distribution' with 'macOS version'.

I claim this edit correctly removes a reference to Linux from the macOS 'Uninstall .NET' section. However, I do not know whether 

'These directories may vary depending on your macOS version'

is true. 

I open the pull request and request a review of this change. 

It may be that the better edit is to remove the whole sentence 
'These directories may vary depending on your Linux distribution.'