Edit publishing-with-visual-studio-code.md (#39285)
* Edit publishing-with-visual-studio-code.md

This commit edits 
```
HelloWorld.exe (HelloWorld on Linux, not created on macOS.)
```
-> 
```
HelloWorld.exe (HelloWorld on Linux or macOS.)
```

When I followed the steps in this tutorial, 
```bash
$ dotnet publish --configuration Release
```
yielded an executable file with the name of the project directory / namespace. 

`publish` created an executable file; that contradicts
> not created on macOS

So, this commit removes the note that `publish` does not create this file on macOS. 

===========
Environment where I observe this
```
$ dotnet --version
8.0.101
```
macOS Big Sur (version 11.7.10)
============

The version of the text before this commit (`not created on macOS`) may be true for some environments. I make this pull request to signal that it is not true in all environments.

* Update publishing-with-visual-studio-code.md

Suppose that the commit before this is correct. That is, suppose that the docs should be edited from
```
HelloWorld.exe (HelloWorld on Linux, not created on macOS.)
```
to
```
HelloWorld.exe (HelloWorld on Linux or macOS.)
```

Then the docs should also be edited from 

```
On Windows or Linux, run the app by using the executable.

On Windows, enter .\HelloWorld.exe and press Enter.

On Linux, enter ./HelloWorld and press Enter.
```
to
```
Run the app by using the executable.

On Windows, enter .\HelloWorld.exe and press Enter.

On Linux or macOS, enter ./HelloWorld and press Enter.
```