More details on collection expression conversions (#39581)
* More details on collection expression conversions

Fixes #39157

Add examples on how the compiler converts a collection expression to an underlying collection.

The specification gives the compiler latitude about how to implement any conversion from a collection expression to a sequence. For that reason, I used a couple examples rather than write rules based on the current implementation.

* lint

* respond to review feedback

Lead with the compiler's behavior, then justify it from the spec.