Simplify the .NET breadcrumb config (#40310)
Overall, the Learn sites are moving to a simpler, shorter use of breadcrumbs.

Ideally, We'd have nothing but `Learn / .NET`, but we have hub pages for some areas of .NET:

- C#
- .NET on Azure
- .NET Aspire
- .NET Orleans
- Architecture guides

In those areas, I have three levels. However, the architecture guides are long-form sections, so those have 4 levels to represent the different architecture guides.