Update dotnet-counters.md (#41023)
Clarify that dotnet-counters works with .NET 5+ runtimes, not older framework versions.

Previously individual parts of the doc specified .NET 5 but it was ambiguous everywhere else what runtime was required. I've removed the section by section annotations and replaced it with a versioning statement at the very top that covers the entire doc.

Technically parts of dotnet-counters do work with the .NET Core 3.1 runtime but given that version is out of support it didn't seem worthwhile to complicate the text detailing which features work in 3.1 vs 5.